started modifying dataloading
